Add CRC validation for fineoffset
Data from sensors could frequently come corrupted (on some days I get
tens of broken items from a couple of sensors). And CRC is currently
ignored in telldus-core, so CRC protocol was reverse-engineered using
reveng and hundreds of samples.

Also did small fixes and clean-ups: condition for humidity is
fixed (100% is the limit), and id is renamed (reserved identifier).

Signed-off-by: Denis Zalevskiy <denis@visfun.org>